fix(elf): compute program-header table size in 64-bit (CodeQL cpp/integer-multiplication-cast-to-long)
elf_load_image() bounds-checks the program headers with
  hdr->e_phoff + hdr->e_phnum * hdr->e_phentsize > size
e_phnum and e_phentsize are uint16_t, so they promote to int and the
product is computed in 32-bit. A crafted header with large values (up to
65535 * 65535 = 0xFFFE0001) overflows int — signed overflow is undefined
behaviour, and a wrapped-small result could slip a too-large table past
this guard before it is widened to the 64-bit comparison.

CodeQL flags this as cpp/integer-multiplication-cast-to-long (CWE-190).
Cast e_phnum to uint64_t so the size is computed exactly in 64-bit.

+++ kernel/elf.c
@@ -24,7 +24,13 @@ int elf_load_image(const uint8_t* data, uint32_t size, uint64_t** out_pd,
     if (!elf_validate(data, size)) return -1;
 
     elf64_hdr_t* hdr = (elf64_hdr_t*)data;
-    if (hdr->e_phoff + hdr->e_phnum * hdr->e_phentsize > size) return -1;
+    // e_phnum and e_phentsize are uint16_t, so they promote to int and the
+    // program-header table size is computed in 32-bit — for hostile values
+    // (up to 65535 * 65535) that overflows int (signed UB) before it widens to
+    // the 64-bit comparison, which could sneak a too-large table past this bound
+    // check (CodeQL cpp/integer-multiplication-cast-to-long). Do the math in
+    // 64-bit so the size is exact.
+    if (hdr->e_phoff + (uint64_t)hdr->e_phnum * hdr->e_phentsize > size) return -1;
 
     uint64_t* pd = alloc_page_directory();
     if (!pd) return -1;
